requirements.txt
# Google Agent Development Kit
google-adk>=1.0.0

# Google AI/Gemini
google-genai>=0.5.0

# A2A Protocol
a2a-python>=0.3.0

# MCP Support
mcp>=1.0.0

# Web Framework for deployment
fastapi>=0.100.0
uvicorn>=0.23.0

# Observability
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Financial calculations
yfinance>=0.2.0

# Validation
pydantic>=2.0.0

# Testing and evaluation
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.24.0
aiohttp>=3.8.0

# Logging
structlog>=23.1.0
